menu.js: Trigger 'run' event for each block in script
We define a `run` function which is responsible
for running (well, triggering the 'run' event actually)
all blocks in the script.

This function is passed to `requestAnimationFrame`, which
is used by the browser for animation. It takes a function
which will be called for the next frame to be rendered by the
browser.

NB: A key function tasked with executing blocks.